### 2019-11-17 #1899 @whedon
[PRE REVIEW]: c14bazAAR: An R package for downloading and preparing C14 dates from different source databases
**Submitting author:** @nevrome (<a href="http://orcid.org/0000-0003-3448-5715">Clemens Schmid</a>)
**Repository:** <a href="https://github.com/ropensci/c14bazAAR" target ="_blank">https://github.com/ropensci/c14bazAAR</a>
**Version:** 1.1.0
