### 2019-11-17 #1899 @labarba
OK, I have found it. You provided the archive DOI (https://doi.org/10.17605/OSF.IO/3DS6A) pointing to your OSF project, but we needed the URL of the GitHub repo: https://github.com/ropensci/c14bazAAR

Has this submission gone through the rOpenSci review?
