### 2019-11-17 #1899 @nevrome
@labarba Thank you for working with this submission!

It did indeed go through the ROpenSci review: https://github.com/ropensci/software-review/issues/333

There I was [specifically asked](https://github.com/ropensci/software-review/issues/333#issuecomment-531599658) to submit with the archive DOI:

> Submit to JOSS using the Zenodo DOI. They will tag it for expedited review.

Maybe they should update their template?
