### 2019-11-17 #1899 @labarba
@arfon Can you clarify this issue about submissions coming from rOpenSci?
