### 2019-11-18 #1899 @arfon
> @arfon Can you clarify this issue about submissions coming from rOpenSci?

This is incorrect. We can use the Zenodo DOI for the submission once we're ready to accept, but we need authors to submit to JOSS using the standard repository address (https://github.com/ropensci/c14bazAAR).

I've asked for guidance on how to fix this in https://github.com/ropensci/software-review/issues/333#issuecomment-554874000
