### 2019-11-18 #1899 @nevrome
@labarba What's the procedure to apply some small changes to the proof? Resubmission?
