### 2019-11-18 #1899 @arfon
@nevrome - please just make the changes to your `paper.md` and ask Whedon to recompile here with `@whedon generate pdf`
