### 2019-11-24 #1899 @nevrome
@labarba The text is fine now. What are the next steps?
