### 2019-11-24 #1914 @whedon
[REVIEW]: c14bazAAR: An R package for downloading and preparing C14 dates from different source databases
**Submitting author:** @nevrome (<a href="http://orcid.org/0000-0003-3448-5715">Clemens Schmid</a>)
**Repository:** <a href="https://github.com/ropensci/c14bazAAR" target ="_blank">https://github.com/ropensci/c14bazAAR</a>
**Version:** 1.1.0
**Archive:** <a href="https://doi.org/10.17605/OSF.IO/3DS6A" target="_blank">10.17605/OSF.IO/3DS6A</a>
