### 2019-11-24 #1914 @arfon
Accepting because of https://github.com/ropensci/software-review/issues/333
